Extract helper functions into separate module
