+++ tests/conftest.py
@@ -1,0 +1,5 @@
+"""Global fixtures and setup code for the test suite."""
+import sys
+import pathlib
+
+sys.path.append(str(pathlib.Path(__file__).parent / "helpers"))
